backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0

# Redis (caching, rate limiting, sessions)
redis==5.0.1
hiredis==2.2.3

# Background Tasks
celery==5.3.4
flower==2.0.1

# Email
aiosmtplib==3.0.1

# Payment Gateways
stripe==7.8.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-slugify==8.0.1
pytz==2023.3
